Projects/WIP/session4_redis/requirements.txt
Pillow==10.4.0
redis==5.0.1

